+++ packages/action-sheet/src/index.ts
@@ -1,6 +1,2 @@
-export * from './action-sheet'
-export * from './action-sheet-item'
-export * from './action-sheet-select'
-export * from './action-sheet-select-button'
-export * from './action-sheet-select-option'
-export * from './action-sheet-select-options'
+export * from '@titicaca/tds-ui/src/components/action-sheet'
+export * from '@titicaca/tds-ui/src/components/action-sheet-select'
