+++ packages/tds-ui/src/components/action-sheet-select/action-sheet-select-options.tsx
@@ -5,9 +5,13 @@ import {
 } from '@floating-ui/react'
 import { FlexBox } from '@titicaca/core-elements'
 
+import { ActionSheetOverlay } from '../action-sheet/action-sheet-overlay'
+import {
+  ActionSheetBody,
+  ActionSheetBodyProps,
+} from '../action-sheet/action-sheet-body'
+
 import { useActionSheetSelect } from './action-sheet-select-context'
-import { ActionSheetOverlay } from './action-sheet-overlay'
-import { ActionSheetBody, ActionSheetBodyProps } from './action-sheet-body'
 
 export type ActionSheetSelectOptions = Pick<
   ActionSheetBodyProps,
