+++ packages/tds-ui/src/components/action-sheet/index.ts
@@ -1,0 +1,3 @@
+export * from './action-sheet-body'
+export * from './action-sheet-item'
+export * from './action-sheet'
